requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic-settings
alembic
python-jose[cryptography]
passlib[bcrypt]
reportlab
pytest
pytest-asyncio 
httpx
